Android instructions fixed. Numbers instead of bullets.
In chrome for some reason on the NPM page https://www.npmjs.com/package/react-native-bluetooth-status - it changed the numbering to 1, then two bullets then 1 and 2. 

When I looked at the file the way it was before my fix, everything here on github seems OK and seemingly there is no need for change. 
So I thought this could be a historical mistake from an older version, where the bullets were not indented, but history shows that the two bullets have been inserted only once and never changed.   So I propose this fix, which explicitely gives two sub numbers. 

I also added an example of what the addition in java should look like.